requirements.txt
fastapi~=0.123.4
uvicorn==0.30.1
aiogram==3.6.0
pydantic-settings==2.3.4
openai>=1.56.0
httpx>=0.28.0
gspread==6.1.2
google-auth==2.30.0
